moving `sanitize_doc` and `verify_filled` functions
